requirements.txt
altair==5.1.2
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.3
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
gitdb==4.0.10
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.24.4
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
streamlit
lightgbm
evidently